scan-irb: Add staff and swab variables to Vietnamese
Add the staff observed and swab type variables to the Vietnamese
translation of the results template. Note that this modifies the second
instance of the disclaimer about self-collected swabs by standardizing
the two disclaimer paragraphs (one towards the top of the page and one
towards the bottom) to contain the exact same text.